add sample_reset_all workflow
